Fix indexing bug in `releasedata_06` which reported wrong release button.
How to reproduce
----------------
Use this version of ctmouse, with a serial microsoft protocol mouse (no wheel)
in COM1, start it with `ctmouse /S1`.
In Monkey Island 1 or 2:
- Hover mouse over an active object that can be "Looked at" (e.g. the
  governor's poster on the first screen of Monkey Island 1).
- Push Left mouse button without releasing, this triggers "Walk to poster"
- Relase Left mouse button. This triggers the right-click "Look at" verb,
  instead of "Walk to".
This suggests that release button data is wrong.

Explanation of the fix
----------------------
The bug seems to be in the int33,6 routine `releasedata_06`.
Register `bx` is the button number:
0 for Left, 1 for Right, 2 for Middle, and -1 for Wheel.
Within the function `bx` gets incremented so its values are:
0 for Wheel, 1 for Left, 2 for Right, and 3 for Middle.
Register `bx` is used to index array `buttrelease` each element of which
is BUTTLASTSTATE size long, which is 6 bytes.
So we would expect the values of `bx` to be {0, 6, 9, 12} but instead they are
{0, 3, 6, 9}.

This seems to fix the issue, though I am not sure how the original code worked.
The original build of ctmouse does not have this issue.